# services/revenue_api/openapi.yaml
openapi: 3.0.3
info:
  title: Revenue API (Test)
  version: '1.0.0'
  description: Email-attributed revenue via Klaviyo Campaign+Flow reports
servers:
  - url: http://localhost:9090
paths:
...
  /clients/{client_id}/revenue/last7:
    get:
      summary: Get last 7-day email-attributed revenue for a client
      parameters:
        - name: client_id
          in: path
          required: true
          schema:
            type: string
        - name: timeframe_key
          in: query
          required: false
          schema:
            type: string
            enum: [last_7_days, last_30_days, last_90_days, yesterday, last_24_hours]
        - name: tz
          in: query
          required: false
          schema:
            type: string
        - name: start
          in: query
          required: false
          schema:
            type: string
          description: ISO8601 start (use with end)
        - name: end
          in: query
          required: false
          schema:
            type: string
          description: ISO8601 end (use with start)
        - name: timeframe
          in: query
          required: false
          schema:
            type: string
          description: URL-encoded JSON timeframe object {"key":"..."} or {"start":"...","end":"..."}
        - name: recompute
          in: query
          required: false
          schema:
            type: boolean
        - name: metric_name
          in: query
          required: false
          schema:
            type: string
          description: Metric name (default 'Placed Order'). metric_id is resolved from Firestore for the client.
      responses:
        '200':
          description: Revenue result
          content:
            application/json:
              schema:
                type: object
                properties:
                  client_id:
                    type: string
                  metric_id:
                    type: string
                  campaign_total:
                    type: number
                    format: float
                  flow_total:
                    type: number
                    format: float
                  total:
                    type: number
                    format: float
                  timeframe:
                    type: string
...